fix : fullScreenByDefault is unreliable
On some devices, the `VideoPlayerController` can be ready before the `Chewie` widget is.
If that happens, the `enterFullScreen()` call does not work and the controller is out of sync believing it’s in full screen where it’s not.

Fix attempt : We wait until we have at least one listener before trying to enter full screen